Correctly handle RenderInstances for subgraph datums
